Correctly update game display on window resize

+++ src/gui/Game.cpp
@@ -16,7 +16,7 @@ Gui::Game::~Game() {
   }
 }
 
-void Gui::Game::updateSize(uint32_t width, uint32_t height) {
+void Gui::Game::updateSize(uint16_t width, uint16_t height) {
   _width = width;
   _height = height;
   if (_surface)
